requirements.txt
python-dotenv
pandas
sqlalchemy
bs4
spacy
undetected-chromedriver
lxml